% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.ellipse.R
\name{all.shape.dists}
\alias{all.shape.dists}
\title{Return all shapes and distances for each zone}
\usage{
\method{all}{shape.dists}(s, na, coords)
}
\description{
Return all shapes and distances for each zone
}
\keyword{internal}
