% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.ellipse.R
\name{all_shape_dists}
\alias{all_shape_dists}
\title{Return all shapes and distances for each zone}
\usage{
all_shape_dists(s, na, coords)
}
\description{
Return all shapes and distances for each zone
}
\keyword{internal}
