% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check_bn_test.R
\name{arg_check_bn_test}
\alias{arg_check_bn_test}
\title{Argument checking for bn.test}
\usage{
arg_check_bn_test(coords, cases, pop, cstar, longlat, alpha, noc)
}
\arguments{
\item{coords}{An \eqn{n \times 2} matrix of centroid
coordinates for the regions.}

\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each
region.}

\item{cstar}{A non-negative integer indicating the
minimum number of cases to include in each window.}

\item{longlat}{The default is \code{FALSE}, which
specifies that Euclidean distance should be used.If
\code{longlat} is \code{TRUE}, then the great circle
distance is used to calculate the intercentroid
distance.}

\item{alpha}{The significance level to determine whether
a cluster is signficant.  Default is 0.10.}

\item{noc}{A logical value indicating whether all
significant clusters should be returned (\code{FALSE})
or only the non-overlapping clusters (\code{TRUE})
arranged in order of significance.  The default is
\code{TRUE}.}

\item{nsim}{The number of simulations from which to
compute the p-value.}

\item{modified}{A logical value indicating whether a
modified version of the test should be performed.  The
original paper recommends computing the p-value for
each cluster as \code{1 - ppois(cstar - 1, lambda =
expected)}. The modified version replaces \code{cstar}
with \code{cases}, the observed number of cases in the
region, and computes the p-value for the cluster as
\code{1 - ppois(cases - 1, lambda = ex)}. The default
is \code{modified = FALSE}.}
}
\description{
Check the arguments of the bn.test function
}
\keyword{internal}
