% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.ellipse.R
\name{arg_check_dist_ellipse}
\alias{arg_check_dist_ellipse}
\title{Check argments if dist.ellipse}
\usage{
arg_check_dist_ellipse(coords, shape, angle)
}
\description{
Check argments if dist.ellipse
}
\keyword{internal}
