% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_arg_check_dweights.R
\name{arg_check_dweights}
\alias{arg_check_dweights}
\title{Argument checking for dweights/tango.weights}
\usage{
arg_check_dweights(coords, kappa, longlat, type, pop)
}
\arguments{
\item{coords}{Matrix of coordinates}

\item{kappa}{Scale parameter for weight function}

\item{longlat}{Logical value indicating whether distance
should be Euclidean (FALSE) or great citcle (TRUE)}

\item{type}{Type of weights (basic, rogerson, or tango)}

\item{pop}{vector of population values}
}
\description{
Argument checking for dweights/tango.weights
}
\keyword{internal}
