% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflex.zones.R
\name{arg_check_rflex_zones}
\alias{arg_check_rflex_zones}
\title{Check arguments of rflex.zones}
\usage{
arg_check_rflex_zones(nn, w, cases, ex, alpha1, type, pop, progress,
  verbose)
}
\description{
Check the arguments of rflex.zones
}
\keyword{internal}
