% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check_scan_test.R
\name{arg_check_scan_test}
\alias{arg_check_scan_test}
\title{Argument checking for many scan-like functions}
\usage{
arg_check_scan_test(coords, cases, pop, ex, nsim, alpha, nreport, ubpop,
  longlat, parallel, k, w, type = NULL)
}
\description{
Check the arguments of many \code{*.test} functions.
}
\keyword{internal}
