% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_arg_check_scan_zones.R
\name{arg_check_scan_zones}
\alias{arg_check_scan_zones}
\title{Argument checking for scan.zones}
\usage{
arg_check_scan_zones(coords, pop, ubpop, longlat)
}
\description{
Check the arguments of the scan.zones function
}
\keyword{internal}
