% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_arg_check_sim.R
\name{arg_check_sim}
\alias{arg_check_sim}
\title{Argument checking for *.sim functions}
\usage{
arg_check_sim(
  nsim,
  ty,
  ex,
  type,
  nn = NULL,
  zones = NULL,
  ein = NULL,
  eout = NULL,
  tpop = NULL,
  popin = NULL,
  popout = NULL,
  w = NULL,
  pop = NULL,
  ubpop = NULL,
  static,
  simdist = "multinomial"
)
}
\description{
Check the arguments of the \code{*.sim} functions.
}
\keyword{internal}
