% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tango.test.R
\name{arg_check_tango_test}
\alias{arg_check_tango_test}
\title{Argument checking for tango.test}
\usage{
arg_check_tango_test(cases, pop, w, nsim)
}
\description{
Check the arguments of the tango.test function
}
\keyword{internal}
