% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.zones.R
\name{combine.zones}
\alias{combine.zones}
\title{Combine distinct zones}
\usage{
combine.zones(z1, z2)
}
\arguments{
\item{z1}{A list of zones}

\item{z2}{A list of zones}
}
\value{
A list of distinct zones
}
\description{
\code{combine.zones} combines the elements of \code{z1}
and \code{z2} into a single list, returning only the
unique zones.
}
\examples{
z1 = list(1:2, 1:3)
z2 = list(2:1, 1:4)
combine.zones(z1, z2)
}
