% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.zones.R
\name{dc.zones}
\alias{dc.zones}
\title{Determine zones for the Double Connection scan test}
\usage{
dc.zones(coords, cases, pop, w, ex = sum(cases)/sum(pop) * pop,
  ubpop = 0.5, ubd = 1, longlat = FALSE, type = "maxonly",
  cl = NULL, progress = TRUE)
}
\arguments{
\item{coords}{An \eqn{n \times 2} matrix of centroid
coordinates for the regions.}

\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each
region.}

\item{w}{A binary spatial adjacency matrix for the regions.}

\item{ex}{The expected number of cases for each region.
The default is calculated under the constant risk
hypothesis.}

\item{ubpop}{The upperbound of the proportion of the
total population to consider for a cluster.}

\item{ubd}{The upperbound for the radius of a cluster. 
This should be a proportion in (0, 1].  The value is
the proportion of the maximum intercentroid distance
between any two locations in \code{coords}. See
Details.}

\item{longlat}{The default is \code{FALSE}, which
specifies that Euclidean distance should be used.If
\code{longlat} is \code{TRUE}, then the great circle
distance is used to calculate the intercentroid
distance.}

\item{type}{One of \code{"maxonly"}, \code{"pruned"}, or 
\code{"all"}.  See Details.}

\item{cl}{
A cluster object created by \code{\link[parallel]{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations.
}

\item{progress}{A logical value indicating whether a progress bar should be displayed.  The default is \code{TRUE}.}
}
\value{
Returns a list of relevant information.  See 
  Details.
}
\description{
\code{dc.zones} determines the zones for the Double 
Connection scan test (\code{\link{dc.test}}).
}
\details{
Every zone considered must have a total population less
than \code{ubpop * sum(pop)}.  Additionally, the maximum
intercentroid distance for the regions within a zone must
be no more than \code{ubd * the maximum intercentroid
distance across all regions}.

\code{type} is a character vector indicating what should 
be returned by the function.  If \code{type = "maxonly"},
then the maximum test statistic from each starting region
is returned .  If \code{type = "pruned"}, the function
returns a list that includes the location ids, test
statistic, total cases, expected cases, and total
population for the zone with the maximum test statistic
for each starting region.  If \code{type = "all"}, the
function returns a list of lists that includes the
location ids, test statistic, total cases, expected
cases, and total population for the sequence of candidate
zones associated with each starting region.
}
\examples{
data(nydf)
data(nyw)
coords = as.matrix(nydf[,c("longitude", "latitude")])
# find zone with max statistic starting from each individual region
max_zones = dc.zones(coords, cases = floor(nydf$cases),
                     nydf$pop, w = nyw, ubpop = 0.25,
                     ubd = .25, longlat = TRUE)
}
\references{
Costa, M.A. and Assuncao, R.M. and Kulldorff, M. (2012)
  Constrained spanning tree algorithms for
  irregularly-shaped spatial clustering, Computational
  Statistics & Data Analysis, 56(6), 1771-1783. 
  <https://doi.org/10.1016/j.csda.2011.11.001>
}
\author{
Joshua French
}
