% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic.penalty.R
\name{elliptic.penalty}
\alias{elliptic.penalty}
\title{Compute elliptic penalty}
\usage{
elliptic.penalty(a, shape)
}
\arguments{
\item{a}{Penalty scale}

\item{shape}{Shape of ellipse.}
}
\value{
A vector of penalities
}
\description{
Compute eccentricity penalty for elliptic scan method.
}
\examples{
elliptic.penalty(a = 0.5, shape = c(1, 1.5, 2))
}
