% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic.test.R
\name{elliptic.test}
\alias{elliptic.test}
\title{Elliptical Spatial Scan Test}
\usage{
elliptic.test(coords, cases, pop, ex = sum(cases)/sum(pop) * pop,
  nsim = 499, alpha = 0.1, ubpop = 0.5, shape = c(1, 1.5, 2, 3, 4,
  5), nangle = c(1, 4, 6, 9, 12, 15), a = 0.5, cl = NULL,
  type = "poisson")
}
\arguments{
\item{coords}{An \eqn{n \times 2} matrix of centroid coordinates for the regions.}

\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each region.}

\item{ex}{The expected number of cases for each region.  The default is calculated under the constant risk hypothesis.}

\item{nsim}{The number of simulations from which to compute the p-value.}

\item{alpha}{The significance level to determine whether a cluster is signficant.  Default is \code{0.10}.}

\item{ubpop}{The upperbound of the proportion of the total population to consider for a cluster.}

\item{shape}{The ratios of the major and minor axes of the 
desired ellipses.}

\item{nangle}{The number of angles (between 0 and 180) to consider for
each shape.}

\item{a}{The penalty for the spatial scan statistic.  The default is 0.5.}

\item{cl}{
A cluster object created by \code{\link[parallel]{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations.
}

\item{type}{The type of scan statistic to implement.  Default is \code{"poisson"}.  Only \code{"poisson"} is currently implemented.}
}
\value{
Returns a list of length two of class scan. The first element (clusters) is a list containing the significant, non-overlappering clusters, and has the the following components:
\item{locids}{The location ids of regions in a significant cluster.} 
\item{coords}{The centroid of the significant clusters.}
\item{r}{The radius of the cluster (the largest intercentroid distance for regions in the cluster).}
\item{pop}{The total population of the regions in the cluster.}
\item{cases}{The observed number of cases in the cluster.}
\item{expected}{The expected number of cases in the cluster.}
\item{smr}{Standarized mortaility ratio (observed/expected) in the cluster.}
\item{rr}{Relative risk in the cluster.}
\item{loglikrat}{The loglikelihood ratio for the cluster (i.e., the log of the test statistic).}
\item{pvalue}{The pvalue of the test statistic associated with the cluster.}
The second element of the list is the centroid coordinates.  This is needed for plotting purposes.
}
\description{
\code{elliptic.test} performs the spatial scan test of Kulldorf (1997).
}
\details{
The test is performed using the spatial scan test based on the Poisson test statistic and a fixed number of cases.  Candidate zones are circular and extend from the observed data locations.  The clusters returned are non-overlapping, ordered from most significant to least significant.  The first cluster is the most  cflikely to be a cluster.  If no significant clusters are found, then the most likely cluster is returned (along with a warning).
}
\examples{
data(nydf)
coords = with(nydf, cbind(longitude, latitude))
out = elliptic.test(coords = coords, 
                   cases = floor(nydf$cases), 
                   pop = nydf$pop, nsim = 49, 
                   alpha = 0.12,
                   shape = 1.5, nangle = 4)
}
\references{
Kulldorff, M. (1997) A spatial scan statistic. Communications in Statistics -- Theory and Methods, 26:1481-1496.
Kulldorff, M., Huang, L., Pickle, L. and Duczmal, L. (2006) An elliptic spatial scan statistic. Statististics in Medicine, 25:3929-3943. doi:10.1002/sim.2490
}
\seealso{
\code{\link{scan.stat}}, \code{\link{plot.scan}}
}
\author{
Joshua French
}
