% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.zones.R
\name{fast.zones}
\alias{fast.zones}
\title{Determine sequence of fast subset scan zones}
\usage{
fast.zones(cases, pop, ubpop = 0.5)
}
\arguments{
\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each
region.}

\item{ubpop}{The upperbound of the proportion of the
total population to consider for a cluster.}
}
\value{
Returns a vector of regions to sequentially and
  cumulatively consider for clustering.
}
\description{
\code{fast.zones} determines the unique zones obtained by
implementing the fast subset scan method of Neill (2012).
}
\details{
The zones returned must have a total population less than
\code{ubpop * sum(pop)} of all regions in the study area.
}
\examples{
data(nydf)
fast.zones(cases = nydf$cases, pop = nydf$population)
}
\references{
Neill, D. B. (2012), Fast subset scan for
  spatial pattern detection. Journal of the Royal
  Statistical Society: Series B (Statistical
  Methodology), 74: 337-360.
  <doi:10.1111/j.1467-9868.2011.01014.x>
}
\author{
Joshua French
}
