% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_zones.R
\name{logical2zones}
\alias{logical2zones}
\title{Convert logical vector to zone}
\usage{
logical2zones(czones, nn, idx = seq_along(nn))
}
\arguments{
\item{czones}{List of list of logical vectors}

\item{nn}{List of nearest neighbor indices}

\item{idx}{Relevant nn indices}
}
\value{
List of list of zones
}
\description{
Convert logical vector to zone
}
\keyword{internal}
