% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlf_zones_internal.R
\name{mlf_zones_internal}
\alias{mlf_zones_internal}
\title{Determine sequence constrained greedy zones.}
\usage{
mlf_zones_internal(uz, max_neighbors, cases, pop, w, e, ty, max_pop, loglikrat,
  yin, ein, popin, maxonly = TRUE)
}
\arguments{
\item{uz}{A list of unique zones from the algorithm.  The first element is the most likely cluster for a single region.}

\item{max_neighbors}{A list.  The first element has the set of potential regions to add to the zone for uz[[1]].}

\item{cases}{The number of cases in each region.}

\item{pop}{The population size of each region.}

\item{w}{The binary spatial adjacency matrix.}

\item{e}{The expected number of cases for each region.}

\item{ty}{The total number of cases.}

\item{max_pop}{The population upperbound for a candidate zone.}

\item{loglikrat}{is a vector to store the test statistic associated with each element of uz.  The first element already has the test statistic of uz[[1]]}

\item{yin}{is a vector to store the cases in each element of uz.  The first element already has the number of cases for uz[[1]]}

\item{ein}{is a vector to store the expected number of cases associated with each element of uz.  The first element already has the expected cases for uz[[1]]}

\item{popin}{s a vector to store the total population associated with each element of uz.  The first element already has the total population of uz[[1]]}

\item{maxonly}{A logical value indicating whether only the maximum of the log likelihood ratio test statistic across all zones should be returned or more detailed information that includes the location id of each zone and the associated test statistic, counts, expected counts, and population in the zone.}
}
\value{
Returns a list of zones to consider for clustering that includes the location id of each zone and the associated test statistic, counts, expected counts, and population in the zone. Or the largest test statistic if maxonly = TRUE.
}
\description{
\code{mlf_zones_internal} finds the most likely zone using the maxima likelihood first scan algorithm.  It is an internal function meant to do the heavy lifting for the smst.zones function.
}
\author{
Joshua French
}

