% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noc_enn.R
\name{noc_enn}
\alias{noc_enn}
\title{Returned ordered non-overlapping clusters}
\usage{
noc_enn(nn, tobs_nn, shape_nn, angle_nn, nnoc = length(nn))
}
\arguments{
\item{nn}{A nearest neighbor list.}

\item{tobs_nn}{The test statistics for each sequence
candidate zone.}

\item{nnoc}{The number of non-overlapping clusters to
return.}
}
\value{
A list with \code{clusts}, a list with elements
  containing the non-overlapping clusters, and
  \code{tobs}, the associated test statistic for each
  cluster.
}
\description{
\code{noc_nn} returns the non-overlapping clusters
ordered by the values on \code{tobs_nn}.
}
\keyword{internal}
