% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nypoly.R
\docType{data}
\name{nypoly}
\alias{nypoly}
\title{SpatialPolygonsDataFrame for New York leukemia data.}
\format{
A SpatialPolygonDataFrame
}
\source{
Bivand, R. S., Pebesma, E. J., Gomez-Rubio, V.,
and Pebesma, E. J. (2013). Applied Spatial Data Analysis
with R, 2nd edition. New York: Springer.
}
\usage{
data(nypoly)
}
\description{
A SpatialPolygonsDataFrame for the New York
leukemia data in \code{nydf}.  Note that the coordinates
in the polygon have been projected to a different
coordinate system (UTM, zone 18), but the order of the
regions/polygons is the same as in \code{nydf}.
}
