% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyw.R
\docType{data}
\name{nyw}
\alias{nyw}
\title{Adjacency matrix for New York leukemia data.}
\format{
A matrix of dimension 281 x 281.
}
\source{
Waller, L.A. and Gotway, C.A. (2005).  Applied
Spatial Statistics for Public Health Data.  Hoboken, NJ:
Wiley.
}
\usage{
data(nyw)
}
\description{
This data set contains a 281 x 281 adjacency
matrix for the New York leukemia data in \code{nydf}.
}
\references{
Waller, L.A., Turnbull, B.W., Clark, L.C.,
and Nasca, P. (1994) "Spatial Pattern Analysis to Detect
Rare Disease Clusters" in Case Studies in Biometry,
N. Lange, L. Ryan, L. Billard, D. Brillinger, L. Conquest,
and J. Greenhouse (eds.) New York: John Wiley and Sons.
}
