% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.results.R
\name{prep.scan2}
\alias{prep.scan2}
\title{a simplified prep.scan for non-scan methods
orders on p-value instead of test statistic}
\usage{
prep.scan2(tobs, zones, pvalue, coords, cases, pop, ex, longlat,
  w = NULL, d = NULL)
}
\description{
a simplified prep.scan for non-scan methods
orders on p-value instead of test statistic
}
\keyword{internal}
