% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes100k.R
\docType{data}
\name{primes100k}
\alias{primes100k}
\title{A vector of prime numbers}
\format{
An integer vector.
}
\source{
The data were adapted from information at Project
  Gutenburg, which can be accessed at
  \url{https://www.gutenberg.org/dirs/6/65/65.txt}.
}
\usage{
data(primes100k)
}
\description{
\code{primes100k} is an integer vector with
the first 101,341 prime numbers.
}
