% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.smerc_cluster.R
\name{print.smerc_cluster}
\alias{print.smerc_cluster}
\title{Print object of class \code{smerc_cluster}.}
\usage{
\method{print}{smerc_cluster}(x, ...)
}
\arguments{
\item{x}{An object of class \code{smerc_cluster}.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}
}
\description{
Print \code{scan} object
}
\examples{
data(nydf)
coords = with(nydf, cbind(x, y))
out = scan.test(coords = coords, cases = floor(nydf$cases),
                pop = nydf$pop, nsim = 49,
                longlat = TRUE, alpha = 0.12)
out
}
