% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_ubpop-print.R
\name{print.smerc_optimal_ubpop}
\alias{print.smerc_optimal_ubpop}
\title{Print object of class \code{smerc_optimal_ubpop}.}
\usage{
\method{print}{smerc_optimal_ubpop}(x, ...)
}
\arguments{
\item{x}{An object of class \code{smerc_optimal_ubpop}.}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Print \code{smerc_optimal_ubpop} object
}
\examples{
data(nydf)
coords = with(nydf, cbind(longitude, latitude))
ubpop_stats = optimal_ubpop(coords = coords, cases  = nydf$cases,
                            pop = nydf$pop, nsim = 49,
                            ubpop = seq(0.05, 0.5, by = 0.05))
ubpop_stats
}
