% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerc_similarity_test-print.R
\name{print.smerc_similarity_test}
\alias{print.smerc_similarity_test}
\title{Print object of class \code{smerc_similarity_test}.}
\usage{
\method{print}{smerc_similarity_test}(x, ..., digits = 2)
}
\arguments{
\item{x}{An object of class \code{smerc_similarity+test}.}

\item{\dots}{Not currently implemented.}

\item{digits}{Number of significant digits to print.}
}
\description{
Print a \code{smerc_similarity_test} object. If the \code{crayon} package
is installed, then the results are printed in color.
}
