% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tango-print.R
\name{print.tango}
\alias{print.tango}
\title{Print object of class \code{tango}.}
\usage{
\method{print}{tango}(x, ..., digits = 2)
}
\arguments{
\item{x}{An object of class \code{tango}.}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}

\item{digits}{Number of significant digits to print.}
}
\description{
Print a \code{tango} object. If the \code{crayon} package
is installed, then the results are printed in color.
}
\examples{
data(nydf)
coords = as.matrix(nydf[,c("x", "y")])
w = dweights(coords, kappa = 1)
results = tango.test(nydf$cases, nydf$pop, w, nsim = 49)
results
}
