% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.stat.R
\name{scan.stat}
\alias{scan.stat}
\title{Spatial scan statistic}
\usage{
scan.stat(yin, ein, eout, ty, type = "poisson")
}
\arguments{
\item{yin}{The total number of cases in the zone.}

\item{ein}{The expected number of cases in the zone.
Conventionally, this is the estimated overall 
  disease risk across the study area, multiplied by the 
  total population size of the zone.}

\item{eout}{The expected number of cases outside the 
zone.  This should be \code{ty - ein}, but
this is not done for efficiency purposes.}

\item{ty}{The total number of cases in the study area.}

\item{type}{The type of scan statistic to implement. 
Currently, only \code{"poisson"} is implemented.}
}
\value{
A vector of scan statistics.
}
\description{
\code{scan.stat} calculates the spatial scan statistic 
for a zone (set of spatial regions).
}
\examples{
# statistic for most likely cluster of 
# New York leukemia data
scan.stat(106, 62.13, 552 - 62.13, 552)
}
\references{
Kulldorff, M. (1997) A spatial scan 
  statistic. Communications in Statistics -- Theory and 
  Methods 26, 1481-1496.
}
\author{
Joshua French
}
