% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerc-package.R
\docType{package}
\name{smerc}
\alias{smerc-package}
\alias{smerc}
\title{smerc}
\description{
*S*tatistical *ME*thods for *R*egional *C*ounts
}
\details{
The **smerc** package implements statistical methods for
analyzing the counts of areal data, with a focus on the
detection of spatial clusters and clustering. The package
has a heavy emphasis on spatial scan methods.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jfrench/smerc/issues}
}

}
\author{
\strong{Maintainer}: Joshua French \email{joshua.french@ucdenver.edu} (\href{https://orcid.org/0000-0002-9708-3353}{ORCID})

Other contributors:
\itemize{
  \item Mohammad Meysami (\href{https://orcid.org/0000-0002-3322-5244}{ORCID}) [contributor]
}

}
