% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerc_cluster.R
\name{smerc_cluster}
\alias{smerc_cluster}
\title{Prepare \code{smerc_cluster}}
\usage{
smerc_cluster(
  tobs,
  zones,
  pvalue,
  coords,
  cases,
  pop,
  ex,
  longlat,
  method,
  rel_param,
  w = NULL,
  d = NULL,
  a = NULL,
  shape_all = NULL,
  angle_all = NULL
)
}
\arguments{
\item{tobs}{The vector of observed test statistics for each zone}

\item{zones}{A list of zones}

\item{pvalue}{The p-value associated with each test statistic}

\item{coords}{An \eqn{n \times 2} matrix of centroid
coordinates for the regions.}

\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each
region.}

\item{ex}{The expected number of cases for each region.
The default is calculated under the constant risk
hypothesis.}

\item{longlat}{The default is \code{FALSE}, which
specifies that Euclidean distance should be used. If
\code{longlat} is \code{TRUE}, then the great circle
distance is used to calculate the intercentroid
distance.}

\item{method}{A character string indicating the method
used to construct the \code{smerc_cluster}.}

\item{rel_param}{A names list with the relevant parameters
associated with \code{method}.}

\item{w}{A binary spatial adjacency matrix for the
regions.}

\item{d}{A precomputed distance matrix based on \code{coords}}

\item{a}{A single value >= 0 indicating the penalty to use
for \code{\link{elliptic.test}}.}

\item{shape_all}{A vector of shape parameters associated
with \code{zones}.}

\item{angle_all}{A vector of angle parameter associated with
\code{zones}.}
}
\value{
A \code{smerc_cluster} object
}
\description{
\code{smerc_cluster} prepares a returns a
\code{smerc_cluster}.
}
