% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smst.zones.R
\name{smst.zones}
\alias{smst.zones}
\title{Determine zones using the static minimum spanning tree scan test of Assuncao et al. (2006)}
\usage{
smst.zones(coords, cases, pop, w, ex = sum(cases)/sum(pop) * pop,
  ubpop = 0.5, ubd = 1, lonlat = FALSE, parallel = TRUE, nsim = 0)
}
\arguments{
\item{coords}{\eqn{An n \times 2} matrix of centroid coordinates for the regions.}

\item{cases}{The number of cases in each region.}

\item{pop}{The population size of each region.}

\item{w}{The binary spatial adjacency matrix.}

\item{ex}{The expected number of cases for each region. The default is calculated under the constant risk hypothesis.}

\item{ubpop}{The upperbound of the proportion of the total population to consider for a cluster.}

\item{ubd}{The upperbound of the proportion of the maximum intercentroid distance across all regions.  Default is 1.}

\item{lonlat}{A logical value indicating whether Euclidean (\code{lonlat = FALSE}) or great circle distance (\code{lonlat = TRUE}) should be used to calculate intercentroid distance.  The default is \code{FALSE}.}

\item{parallel}{A logical value indicating whether the algorithm should be parallelized.  Default is \code{TRUE}.}

\item{nsim}{An non-negative integer indicating the number of simulated data sets to generate under the constant risk hypothesis.}
}
\value{
Returns a list of zones to consider for clustering that includes the location id of each zone and the associated test statistic, counts, expected counts, and population in the zone.
}
\description{
\code{smst.zones} determines the most likely cluster zone obtained by implementing the static minimum spanning tree (SMST) scan test of Assuncao et al. (2006)
}
\details{
Each step of the SMST scan test seeks to maximize the likelihood ratio test statistic used in the original spatial scan test (Kulldorff 1995).  The first zone considered is the region that maximizes this statistic, providing that no more than \code{ubpop} proportion of the total population is in the zone.  The second zone is the first zone and the connected region that maximizes the scan statistic, subject to the population and distance constraints.

Every zone considered must have a total population less than \code{ubpop * the total population across all regions} in the study area.  Additionally, the maximum intercentroid distance for the regions within a zone must be no more than \code{ubd * the maximum intercentroid distance across all regions}.

See also the maxima likelihood first scan method of Yao et al. (2011).
}
\examples{
data(nydf)
data(nyw)
coords = as.matrix(nydf[,c("x", "y")])
smst.zones(coords, cases = floor(nydf$cases), nydf$pop, w = nyw, ubpop = 0.1, ubd = .5, nsim = 10)
}
\author{
Joshua French
}
\references{
Assuncao, R.M., Costa, M.A., Tavares, A. and Neto, S.J.F. (2006). Fast detection of arbitrarily shaped disease clusters, Statistics in Medicine, 25, 723-742.
Yao, Z., Tang, J., & Zhan, F. B. (2011). Detection of arbitrarily-shaped clusters using a neighbor-expanding approach: A case study on murine typhus in South Texas. International journal of health geographics, 10(1), 1.
}

