% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scan.R
\name{summary.scan}
\alias{summary.scan}
\title{Summary of \code{scan} object}
\usage{
\method{summary}{scan}(object, ..., idx = seq_along(object$clusters),
  digits = 1)
}
\arguments{
\item{object}{An object of class \code{scan}.}

\item{...}{Arguments passed on to \code{base::summary}
\describe{
  \item{object}{an object for which a summary is desired.}
}}

\item{idx}{A index vector indicating the elements of
\code{object$clusters} to print information for. The default
is all clusters.}

\item{digits}{Integer indicating the number of decimal places.}
}
\description{
Summary of object of class \code{scan}.
}
\examples{
data(nydf)
coords = with(nydf, cbind(x, y))
out = scan.test(coords = coords, cases = floor(nydf$cases),
                pop = nydf$pop, nsim = 49,
                longlat = TRUE, alpha = 0.12)
summary(out)
}
