% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uls.zones.R
\name{uls.zones}
\alias{uls.zones}
\title{Determine sequence of ULS zones.}
\usage{
uls.zones(cases, pop, w, ubpop = 0.5)
}
\arguments{
\item{cases}{The number of cases observed in each region.}

\item{pop}{The population size associated with each
region.}

\item{w}{A binary spatial adjacency matrix for the regions.}

\item{ubpop}{The upperbound of the proportion of the
total population to consider for a cluster.}
}
\value{
Returns a list of zones to consider for
  clustering.  Each element of the list contains a vector
  with the location ids of the regions in that zone.
}
\description{
\code{uls.zones} determines the unique zones obtained by
implementing the ULS (Upper Level Set) test of Patil
and Taillie (2004).  Note: the algorithm implicitly
assumes the estimated rates are all unique.  If this
is not true, the results may differ slightly from the
canonical algorithm proposed by Patil and Taillie (2004).
}
\details{
The zones returned must have a total population less than
\code{ubpop * sum(pop)} of all regions in the study
area.
}
\examples{
data(nydf)
data(nyw)
uls.zones(cases = nydf$cases, pop = nydf$population, w = nyw)
}
\references{
Patil, G.P. & Taillie, C. Environmental and
  Ecological Statistics (2004) 11: 183.
  <doi:10.1023/B:EEST.0000027208.48919.7e>
}
\author{
Joshua French
}
