% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerc_cluster-plot.R
\name{w2segments}
\alias{w2segments}
\title{Returns segments connecting neighbors}
\usage{
w2segments(w, coords)
}
\description{
w2segments takes a binary spatial proximity matrix and
associated coordinates.  It returns segments connecting
the neighbors for plotting purposes
}
\keyword{internal}
