% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.test.R
\name{arg_check_bn_test}
\alias{arg_check_bn_test}
\title{Argument checking for bn.test}
\usage{
arg_check_bn_test(coords, cases, pop, cstar, longlat, alpha, noc)
}
\description{
Check the arguments of the bn.test function
}
\keyword{internal}
