% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic.zones.R
\name{elliptic.zones}
\alias{elliptic.zones}
\title{Determine zones for \code{elliptic.test}}
\usage{
elliptic.zones(coords, pop, ubpop = 0.5, shape = c(1, 1.5, 2, 3, 4, 5),
  nangle = c(1, 4, 6, 9, 12, 15))
}
\value{
A list with all distinct zones, the associated
shape parameter, and the associated angle parameter.
}
\description{
Determine zones for \code{elliptic.test}
}
\examples{
data(nydf)
coords = with(nydf, cbind(longitude, latitude))
out = elliptic.zones(coords = coords, pop = nydf$pop,
                     shape = 1.5, nangle = 4)

}
