% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.zones.R
\name{flex.zones}
\alias{flex.zones}
\title{Determine zones for flexibly shaped spatial scan test}
\usage{
flex.zones(coords, w, k = 10, longlat = FALSE, cl = NULL,
  progress = TRUE, verbose = FALSE)
}
\arguments{
\item{coords}{An \eqn{n \times 2} matrix of centroid
coordinates for the regions.}

\item{w}{A binary spatial adjacency matrix for the regions.}

\item{k}{An integer indicating the maximum number of regions to inclue in a potential cluster.  Default is 10}

\item{longlat}{The default is \code{FALSE}, which
specifies that Euclidean distance should be used.If
\code{longlat} is \code{TRUE}, then the great circle
distance is used to calculate the intercentroid
distance.}

\item{cl}{
A cluster object created by \code{\link[parallel]{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations.
}

\item{progress}{A logical value indicating whether a progress bar should be displayed.  The default is \code{TRUE}.}

\item{verbose}{A logical value indicating whether
descriptive progress messages should be provided.
Default is \code{FALSE}.  If \code{TRUE}, this can be
useful for diagnosing where the sequences of connected
subgraphs is slowing down/having problems.
Additionally, a loop is used to run the algorithm,
rather than \code{\link[base]{lapply}}.  Memory-saving
steps are also taken.}
}
\value{
Returns a list of zones to consider for clustering.  Each element of the list contains a vector with the location ids of the regions in that zone.
}
\description{
\code{flex.zones} determines the unique zones to consider for the flexibly shaped spatial scan test of Tango and Takahashi (2005).  The algorithm uses a breadth-first search to find all subgraphs connected to each vertex (region) in the data set of size \eqn{k} or less.
}
\examples{
data(nydf)
data(nyw)
coords = cbind(nydf$x, nydf$y)
zones = flex.zones(coords, w = nyw, k = 3)
\dontrun{
# see what happens when verbose = TRUE
zones = flex.zones(coords, w = nyw, k = 3, verbose = TRUE)
}
}
\references{
Tango, T., & Takahashi, K. (2005). A flexibly shaped spatial scan statistic for detecting clusters. International journal of health geographics, 4(1), 11.
}
\author{
Joshua French
}
