% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerc_cluster-print.R
\name{print.smerc_cluster}
\alias{print.smerc_cluster}
\title{Print object of class \code{smerc_cluster}.}
\usage{
\method{print}{smerc_cluster}(x, ..., extra = FALSE)
}
\arguments{
\item{x}{An object of class \code{smerc_cluster}.}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}

\item{extra}{A logical value. Default is \code{FALSE}.
\code{TRUE} indicates that extra information should be
printed.}
}
\description{
Print \code{smerc_cluster} object
}
\examples{
data(nydf)
coords = with(nydf, cbind(x, y))
out = scan.test(coords = coords, cases = floor(nydf$cases),
                pop = nydf$pop, nsim = 49,
                longlat = TRUE, alpha = 0.12)
out
}
