% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_ubpop.R
\name{seq_pop_idx}
\alias{seq_pop_idx}
\title{Title}
\usage{
seq_pop_idx(popin, tpop, ubpop_seq)
}
\arguments{
\item{popin}{vector of population in each candidate zone}

\item{tpop}{total population across all regions}

\item{ubpop_seq}{# sequence of population upper bounds to consider}
}
\value{
logical matrix indicating which candidate zones
satisfy the sequence of population constraints
}
\description{
Title
}
