% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uls.zones.R
\name{uls.zones}
\alias{uls.zones}
\title{Determine sequence of ULS zones.}
\usage{
uls.zones(cases, pop, w, ubpop = 0.5)
}
\arguments{
\item{cases}{The number of cases in each region.}

\item{pop}{The population size of each region.}

\item{w}{The binary spatial adjacency matrix.}

\item{ubpop}{The upperbound of the proportion of the total population to consider for a cluster.}
}
\value{
Returns a list of zones to consider for clustering.  Each element of the list contains a vector with the location ids of the regions in that zone.
}
\description{
\code{uls.zones} determines the unique zones obtained by implementing the ULS (Upper Level Set) method of Patil and Taillie (2004).
}
\details{
The zones returned must have a total population less than ubpop * the total population of all regions in the study area.
}
\examples{
data(nydf)
data(nyw)
uls.zones(cases = nydf$cases, pop = nydf$population, w = nyw)
}
\author{
Joshua French
}
\references{
Patil, G. P., and Taillie, C. (2004). Upper level set scan statistic for detecting arbitrarily shaped hotspots. Environmental and Ecological Statistics, 11(2), 183-197.
}

